+++ Sangsoo/study_8th/lib/constants.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+const labelTextStyle = TextStyle(
+    fontSize: 18.0,
+    color:Color(0xFF8D8E98)
+);
+
+const bottomContainerHeight = 80.0;
+const bottomContainerColour = Color(0xFFEB1555);
+const inactiveCardColour = Color(0xFF111328);
+const activeCardColour = Color(0xFF1D1E33);
