+++ Sangsoo/study_8th/lib/input_page.dart
@@ -2,11 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'male_or_female.dart';
 import 'reusable_card.dart';
-
-const bottomContainerHeight = 80.0;
-const bottomContainerColour = Color(0xFFEB1555);
-const inactiveCardColour = Color(0xFF111328);
-const activeCardColour = Color(0xFF1D1E33);
+import 'constants.dart';
 
 enum Gender { male, female }
 
@@ -60,6 +56,10 @@ class _InputPageState extends State<InputPage> {
           Expanded(
             child: ReusableCard(
               colour: activeCardColour,
+              cardChild: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: <Widget>[Text('HEIGHT')],
+              ),
             ),
           ),
           Expanded(
